Update and rename ca-CA.json to ca-ES.json
The language code is not correct. The current one (ca-CA) would be for Catalan spoken in Canada (I wish!), but Catalan is mostly spoken in Spain (along with Andorra, France, and Italy). A correct language code would be ca-ES (although the best practice would be a generic language code "ca.json").

Also, I corrected some typos and untranslated sentences.